Traffic Light Classifier
In this project, you’ll use your knowledge of computer vision techniques to build a classifier for images of traffic lights! You'll be given a datasets of traffic light images in which one of three lights is illuminated: red, yellow, or green.
My code can work perfectly with 100% for all datasets provided by Udacity.